Julia 1.0 updates
