+++ REQUIRE
@@ -2,3 +2,6 @@ julia 0.6
 Flux
 Images
 BSON
+Requires
+REPL
+ColorTypes
