+++ src/vgg19.jl
@@ -41,6 +41,6 @@ VGG19() = VGG19(vgg19_layers())
 
 Base.show(io::IO, ::VGG19) = print(io, "VGG19()")
 
-Flux.treelike(VGG19)
+@treelike VGG19
 
 (m::VGG19)(x) = m.layers(x)
